[deltatocomulativeprocessor]: Move tracking to internal (#5871)
