+++ processor/cumulativetodeltaprocessor/internal/tracking/tracker.go
